Update index.rst
Removed some parts of the structure as they are no longer on the network. Beside that, I can not see the list of the packs in 1.6.4 and 1.7.10 but in 1.6 you might want to remove Monster and Horizons and in 1.7.10 you might want to remove Bevo's Tech pack, Death is expected, MineyourGalaxy? , modsauce, ftb resurrection and the dark trilogy.

Update on skyfactory from skyfactory 2 to skyfactory 2.5

Beside that it might be possible to add 1.8 to the list with skyblock on it as we are having that one as well.