Fix info for "The list of modpacks could not be downloaded" error
The launcher has issues with IPv6 connections, so if an application is forcing java to use IPv6, it will error out with a stacktrace like this: https://i.imgur.com/opjxZXZ.png

This is just some instructions on how to fix it. This should only happen on windows so I only included info for windows. Yet to be reported for linux/mac.

Also fixed some grammatical errors